validate file start
