+++ front-end/src/ReservationForm.jsx
@@ -2,6 +2,8 @@ import React, { useState } from "react";
 import { createReservations } from "./utils/api";
 import { useHistory } from "react-router-dom";
 import ErrorAlert from "./layout/ErrorAlert";
+import { previous } from "./utils/date-time";
+import { mobileValidate } from "./utils/validate";
 export default function ReservationForm() {
 	const history = useHistory();
 	const [formData, setFormData] = useState({
@@ -17,36 +19,8 @@ export default function ReservationForm() {
 		const changeObj = { ...formData };
 		if (event.target.id === `mobile_number`) {
 			let phoneNumber = event.target.value;
-			if (phoneNumber[phoneNumber.length - 1] === "-") {
-				phoneNumber = phoneNumber.slice(0, phoneNumber.length - 1);
-				event.target.value = phoneNumber;
-			}
-			if (phoneNumber.length === 3) {
-				if (formData.mobile_number.length < phoneNumber.length) {
-					phoneNumber += "-";
-					event.target.value = phoneNumber;
-				}
-			}
-
-			if (phoneNumber.length >= 9) {
-				if (!phoneNumber.includes("-")) {
-					const temp1 = phoneNumber.substr(0, 3);
-					const temp2 = phoneNumber.substr(3, 3);
-					const temp3 = phoneNumber.substr(6, 4);
-					phoneNumber = `${temp1}-${temp2}-${temp3}`;
-					event.target.value = phoneNumber;
-				} else if (phoneNumber[7] !== "-") {
-					const temp1 = phoneNumber.substr(0, 3);
-					const temp2 = phoneNumber.substr(4, 3);
-					const temp3 = phoneNumber.substr(7, 4);
-					phoneNumber = `${temp1}-${temp2}-${temp3}`;
-					event.target.value = phoneNumber;
-				}
-			}
-			if (phoneNumber.length > 12) {
-				phoneNumber = phoneNumber.slice(0, 12);
-				event.target.value = phoneNumber;
-			}
+			phoneNumber = mobileValidate(phoneNumber,formData.mobile_number.length);
+			event.target.value = phoneNumber;
 		}
 		changeObj[event.target.id] = event.target.value;
 		changeObj.people = Number(changeObj.people);
@@ -110,6 +84,7 @@ export default function ReservationForm() {
 						onChange={formChange}
 						value={formData.reservation_date}
 						type="date"
+						min={previous(new Date().toISOString().split("T")[0])}
 						placeholder="YYYY-MM-DD"
 						pattern="\d{4}-\d{2}-\d{2}"
 						className="form-control"
